[Feat ✏️] PostDetail Home에 적용
