+++ src/pages/Home/Home.tsx
@@ -10,8 +10,6 @@ import PostContainer from "./components/PostContainer/PostContainer"
 import PostEdit from "@/components/PostEdit/PostEdit"
 import { useNavigate, useParams } from "react-router-dom"
 import usePostEditModalStore from "@/components/PostEdit/stores/usePostEditModalStore"
-import PostDetail from "@/components/PostDetail/PostDetail"
-import usePostDetailModalStore from "@/components/PostDetail/store/usePostDetailModalStore"
 
 const Home = () => {
   // 완료 후 제거 예정
@@ -19,12 +17,9 @@ const Home = () => {
   const { isShowEditModal, showEditModal, closeEditModal } =
     usePostEditModalStore()
 
-  // Post Detail store 상태
-  const { isShowPostDetail, showDetailModal, closeDetailModal } =
-    usePostDetailModalStore()
-
   const { modalName } = useParams()
   const navigate = useNavigate()
+
   const [selectedCategory, setSelectedCategory] =
     useState<Category>(INITIAL_CATEGORY)
 
@@ -36,21 +31,12 @@ const Home = () => {
     if (modalName === "postedit") {
       showEditModal()
     }
-
-    if (modalName === "postdetail") {
-      showDetailModal()
-    }
-  }, [closeEditModal, modalName, showDetailModal, showEditModal])
+  }, [modalName, showEditModal])
 
   const handleClosePostEdit = useCallback(() => {
     closeEditModal()
     navigate("/")
   }, [closeEditModal, navigate])
-
-  const handleClosePostDetail = useCallback(() => {
-    closeDetailModal()
-    navigate("/")
-  }, [closeDetailModal, navigate])
 
   const onSelectedCategory: OnSelectCategory = (newCategory) => {
     setSelectedCategory(newCategory)
@@ -71,14 +57,6 @@ const Home = () => {
           Edit Modal
         </S.TestButton>
 
-        <S.TestButton
-          onClick={() => {
-            navigate("/postdetail/659c181c16a2b736436afca2")
-          }}
-        >
-          Detail Modal
-        </S.TestButton>
-
         <PostContainer selectedCategory={selectedCategory} />
       </S.HomeLayout>
 
@@ -86,13 +64,6 @@ const Home = () => {
         <PostEdit
           onClose={handleClosePostEdit}
           isShowModal={isShowEditModal}
-        />
-      )}
-
-      {isShowPostDetail && (
-        <PostDetail
-          onClose={handleClosePostDetail}
-          isShow={isShowPostDetail}
         />
       )}
     </>
